requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-core
sentence-transformers
faiss-cpu
pypdf
python-dotenv
transformers